[FEAT]: Add useKeyDownEvent custom hook
