+++ src/components/EditorTemplate.js
@@ -4,6 +4,7 @@ import React, { useContext, useEffect, useState } from "react";
 import styled from "styled-components";
 
 import useDragAndDrop from "../hooks/useDragAndDrop";
+import useKeyDownEvent from "../hooks/useKeyDownEvent";
 import useResize from "../hooks/useResize";
 import { Context } from "../store/Store";
 import {
@@ -22,6 +23,17 @@ function EditorTemplate({ editorVersion, userSavedCode }) {
   const [parentRef, targetRef] = useDragAndDrop(isResizing, setIsResizing);
   const { globalState, dispatch } = useContext(Context);
   const { imageData } = globalState;
+
+  useKeyDownEvent(
+    currentEditor,
+    parentRef,
+    targetRef,
+    counter,
+    copyingElement,
+    setCopyingElement,
+    setCurrentEditor,
+    setCounter
+  );
 
   useEffect(() => {
     if (imageData.localImageSrc) {
@@ -68,45 +80,6 @@ function EditorTemplate({ editorVersion, userSavedCode }) {
       dispatch({ type: "SET_EDITOR", payload: parentRef.current });
     }
   }, []);
-
-  useEffect(() => {
-    const handleKeyboardEvent = (event) => {
-      if (event.metaKey && event.key === "z") {
-        if (currentEditor.length >= 2) {
-          currentEditor.pop();
-          parentRef.current.innerHTML = currentEditor[currentEditor.length - 1];
-        }
-      }
-
-      if (event.metaKey && event.key === "c") {
-        if (targetRef.current.tagName !== "DIV") {
-          setCopyingElement(targetRef.current);
-        }
-      }
-      if (event.metaKey && event.key === "v") {
-        const copyingElementLeft = Number(
-          copyingElement.style.left.replace("%", "")
-        );
-        const copyingElementTop = Number(
-          copyingElement.style.top.replace("%", "")
-        );
-        const clonedNode = copyingElement.cloneNode();
-
-        clonedNode.innerHTML = copyingElement.innerHTML;
-        clonedNode.style.left = `${copyingElementLeft + counter}%`;
-        clonedNode.style.top = `${copyingElementTop + counter}%`;
-
-        parentRef.current.appendChild(clonedNode);
-        setCurrentEditor((state) => [...state, parentRef.current.innerHTML]);
-        setCounter((counter) => counter + 1);
-      }
-    };
-
-    document.addEventListener("keydown", handleKeyboardEvent);
-    return () => {
-      document.removeEventListener("keydown", handleKeyboardEvent);
-    };
-  }, [currentEditor, copyingElement, counter]);
 
   return (
     <EditorTemplateBody
