+++ src/services/api.js
@@ -1,5 +1,5 @@
 import axios from "axios";
 
 export default axios.create({
-  baseURL: process.env.REACT_APP_API_ENDPOINT_URL || "http://localhost:8080",
+  baseURL: "http://localhost:8080",
 });
